requirements.txt
requests>=2.31.0
pillow>=10.3.0
matplotlib>=3.9.0
python-dateutil>=2.9.0
pyinstaller>=6.14.1    # Only needed for building executables
